pyproject.toml
[tool.poetry]
name = "exifgpsmapper"
version = "0.1.0"
description = ""
authors = ["Jani Sourander <sourander@gmail.com>"]
readme = "README.md"
packages = [{include = "exif_gps_mapper"}]


[tool.poetry.dependencies]
python = "^3.11"
notebook = "^6.5.2"
pyexiftool = "^0.5.5"
pandas = "^1.5.3"
pyarrow = "^10.0.1"
pyyaml = "^6.0"
flask = "^2.2.2"
requests = "^2.28.2"
gpxpy = "^1.5.0"
folium = "^0.14.0"


[tool.poetry.scripts]
auth = "exif_gps_mapper.authenticate:main"
test = 'project_scripts:test'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
